feat(questions-feed): new layout for group question tiles with fan graphs
